E-paper 이미지 헤더 구조 수정: 데이터 타입 변경 및 오류 처리 추가

+++ apis/send_epaper_data.h
@@ -7,10 +7,10 @@
 // Image header structure matching kernel driver
 typedef struct
 {
-    uint32_t width;
-    uint32_t height;
+    uint16_t width;
+    uint16_t height;
     uint32_t data_length;
-    uint32_t header_checksum;
+    uint16_t header_checksum;
 } __attribute__((packed)) image_header_t;
 
 typedef struct
